--- webui/react/src/components/UserSettings.module.scss
@@ -1,15 +1,0 @@
-.base,
-.field {
-  display: flex;
-  flex-direction: column;
-  
-  .body {
-    color: var(--theme-colors-monochrome-8);
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 8px;
-  }
-  .header {
-    margin-bottom: 8px;
-  }
-}
