+++ webui/react/src/hooks/useModal/UserSettings/useModalUserSettings.module.scss
@@ -1,3 +1,18 @@
 .noFooter :global(.ant-modal-confirm-btns) {
   display: none;
 }
+.base,
+.field {
+  display: flex;
+  flex-direction: column;
+  
+  .body {
+    color: var(--theme-colors-monochrome-8);
+    display: flex;
+    justify-content: space-between;
+    margin-bottom: 8px;
+  }
+  .header {
+    margin-bottom: 8px;
+  }
+}
