+++ webui/react/src/utils/user.test.ts
@@ -1,5 +1,6 @@
-import { getDisplayName } from './user';
 import { DetailedUser } from 'types';
+
+import { getDisplayName } from './user';
 
 const DISPLAY_NAME_VALUE = 'Display Name';
 const USERNAME_VALUE = 'username_1';
